add touristSpot in travel

+++ src/models/toruistspot.js
@@ -13,7 +13,7 @@ module.exports = (sequelize, DataTypes) => {
     latitude: DataTypes.FLOAT,
     days: DataTypes.INTEGER,
     cost: DataTypes.INTEGER,
-    rates: DataTypes.INTEGER,
+    rates: DataTypes.FLOAT,
     transportation: DataTypes.STRING,
     review: DataTypes.STRING,
     photo_url: DataTypes.STRING
